Control.Monad.Trans.AWS

Description

A monad transformer built on top of functions from Network.AWS which encapsulates various common parameters, errors, and usage patterns.

Transformer

data AWST m a

The transformer. This satisfies all of the constraints that the functions in this module require, such as providing MonadResource instances, as well as keeping track of the internal Env environment.

The MonadError instance for this transformer internally uses ExceptT to handle actions that result in an Error. For more information see sendCatch and paginateCatch.

type AWS = AWST IO

A convenient alias for AWST IO.

Run

runAWST :: MonadBaseControl IO m => AWST m a -> Env -> m (Either Error a)

Unwrap an AWST transformer, calling all of the registered ResourceT release actions.

Environment

data Env

The environment containing the parameters required to make AWS requests.

envAuth :: Lens' Env Auth

envRegion :: Lens' Env Region

envManager :: Lens' Env Manager

envLogging :: Lens' Env Logging

Creating the environment

data Credentials

Determines how authentication information is retrieved.

Constructors

FromKeys AccessKey SecretKey
Explicit access and secret keys. Note: you can achieve the same result purely by using fromKeys.

FromSession AccessKey SecretKey SecurityToken
A session containing the access key, secret key, and a security token. Note: you can achieve the same result purely by using fromSession.

FromProfile ByteString
An IAM Profile name to lookup from the local EC2 instance-data.

FromEnv ByteString ByteString
Environment variables to lookup for the access and secret keys.

Discover
Attempt to read the default access and secret keys from the environment, falling back to the first available IAM profile if they are not set.

Note: This attempts to resolve http://instance-data rather than directly retrieving http://169.254.169.254 for IAM profile information to ensure the dns lookup terminates promptly if not running on EC2.

newEnv :: (Functor m, MonadIO m) => Region -> Credentials -> Manager -> ExceptT String m Env

This creates a new environment without debug logging and uses getAuth to expand/discover the supplied Credentials.

Lenses such as envLogging can be used to modify the Env with a debug logger.

Debugging

debug :: (MonadIO m, MonadReader Env m) => Text -> m ()

Use the logger from envLogging to log a debug message.

whenDebug :: MonadReader Env m => m () -> m ()

Perform a monadic action if envLogging is set to Debug.

Analogous to when.

Regionalisation

data Region

The sum of available AWS regions.

Constructors

Ireland
Europe / eu-west-1

Tokyo
Asia Pacific / ap-northeast-1

Singapore
Asia Pacific / ap-southeast-1

Sydney
Asia Pacific / ap-southeast-2

Beijing
China / cn-north-1

NorthVirginia
US / us-east-1

NorthCalifornia
US / us-west-1

Oregon
US / us-west-2

GovCloud
AWS GovCloud / us-gov-west-1

GovCloudFIPS
AWS GovCloud (FIPS 140-2) S3 Only / fips-us-gov-west-1

SaoPaulo
South America / sa-east-1

region :: MonadReader Env m => m Region

Inspect the current scopes' Region.

within :: MonadReader Env m => Region -> m a -> m a

Scope a monadic action within the specific Region.

Helpers

hoistEither :: (MonadError Error m, AWSError e) => Either e a -> m a

Hoist an Either throwing the Left case, and returning the Right.

scoped :: MonadReader Env m => (Env -> m a) -> m a

Pass the current environment to a function.

Requests

Synchronous

send :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSRequest a) => a -> m (Rs a)

Send a data type which is an instance of AWSRequest, returning it's associated Rs response type.

This will throw any HTTPException or AWSServiceError returned by the service using the MonadError instance. In the case of AWST this will cause the internal ExceptT to short-circuit and return an Error in the Left case as the result of the computation.

sendCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSRequest a) => a -> m (Either (Er (Sv a)) (Rs a))

Send a data type which is an instance of AWSRequest, returning either the associated Rs response type in the success case, or the related service's Er type in the error case.

This includes HTTPExceptions, serialisation errors, and any service errors returned as part of the Response.

Paginated

paginate :: (MonadCatch m, MonadResource m, MonadReader Env m, MonadError Error m, AWSPager a) => a -> Source m (Rs a)

Send a data type which is an instance of AWSPager and paginate while there are more results as defined by the related service operation.

Errors will be handle identically to send.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

paginateCatch :: (MonadCatch m, MonadResource m, MonadReader Env m, AWSPager a) => a -> Source m (Either (Er (Sv a)) (Rs a))

Send a data type which is an instance of AWSPager and paginate over the associated Rs response type in the success case, or the related service's Er type in the error case.

Note: The ResumableSource will close when there are no more results or the ResourceT computation is unwrapped. See: runResourceT for more information.

Pre-signing URLs

presign

Arguments

:: (MonadIO m, MonadReader Env m, AWSRequest a, AWSPresigner (Sg (Sv a)))

=> a
Request to presign.

-> UTCTime
Signing time.

-> Int
Expiry time in seconds.

-> m (Signed a (Sg (Sv a)))

Wait for an asynchronous computation initiated by async to complete and raise any returned error case using hoistEither. wait :: (MonadBaseControl IO m, MonadError Error m, AWSError e) => Async (StM m (Either e a)) -> m a wait = Async.wait >=> hoistEither

Presign a URL with expiry to be used at a later time.

Note: Requires the service's signer to be an instance of AWSPresigner. Not all signing process support this.

Asynchronous actions

async :: MonadBaseControl IO m => m a -> m (Async (StM m a))

wait :: MonadBaseControl IO m => Async (StM m a) -> m a

Types

module Network.AWS.Types