Network.AWS.Data

Description

Serialisation classes and primitives for the various formats used to communicate with AWS.

ByteString

type LazyByteString = ByteString

class ToByteString a where

Minimal complete definition

Nothing

Methods

toBS :: a -> ByteString

class ToBuilder a where

Minimal complete definition

Nothing

Methods

build :: a -> Builder

showBS :: ToByteString a => a -> String

buildBS :: ToBuilder a => a -> LazyByteString

stripBS :: ByteString -> ByteString

data Base64

Base64 encoded binary data.

Text

class FromText a where

Methods

parser :: Parser a

fromText :: FromText a => Text -> Either String a

takeLowerText :: Parser Text

matchCI :: Text -> a -> Parser a

class ToText a where

Methods

toText :: a -> Text

showText :: ToText a => a -> String

Numeric

newtype Nat

Constructors

Nat

Fields

unNat :: Natural

_Nat :: Iso' Nat Natural

Time

data Format

Constructors

RFC822Format

ISO8601Format

BasicFormat

AWSFormat

POSIXFormat

data Time :: Format -> * where

Constructors

Time :: UTCTime -> Time a

_Time :: Iso' (Time a) UTCTime

data UTCTime :: *

This is the simplest representation of UTC. It consists of the day number, and a time offset from midnight. Note that if a day has a leap second added to it, it will have 86401 seconds.

type RFC822 = Time RFC822Format

type ISO8601 = Time ISO8601Format

type BasicTime = Time BasicFormat

type AWSTime = Time AWSFormat

type POSIX = Time POSIXFormat

parseTime :: ParseTime a => TimeLocale -> String -> String -> Maybe a

defaultTimeLocale :: TimeLocale

Locale representing American usage.

knownTimeZones contains only the ten time-zones mentioned in RFC 822 sec. 5: "UT", "GMT", "EST", "EDT", "CST", "CDT", "MST", "MDT", "PST", "PDT". Note that the parsing functions will regardless parse single-letter military time-zones and +HHMM format.

iso8601DateFormat :: Maybe String -> String

Construct format string according to ISO-8601.

The Maybe String argument allows to supply an optional time specification. E.g.:

iso8601DateFormat Nothing            == "%Y-%m-%d"           -- i.e. YYYY-MM-DD
iso8601DateFormat (Just "%H:%M:%S")  == "%Y-%m-%dT%H:%M:%S"  -- i.e. YYYY-MM-DDTHH:MM:SS

Sensitive

newtype Sensitive a

read . show /= isomorphic

Constructors

Sensitive

Fields

desensitise :: a

_Sensitive :: Iso' (Sensitive a) a

HTTP

Body

data RsBody

Constructors

RsBody (ResumableSource (ResourceT IO) ByteString)

_RsBody :: Iso' RsBody (ResumableSource (ResourceT IO) ByteString)

connectBody :: MonadResource m => RsBody -> Sink ByteString m a -> m a

data RqBody

Constructors

RqBody

Fields

_bdyHash :: Digest SHA256

_bdyBody :: RequestBody

bdyHash :: Lens' RqBody (Digest SHA256)

bdyBody :: Lens' RqBody RequestBody

bodyHash :: RqBody -> ByteString

isStreaming :: RqBody -> Bool

class ToBody a where

Minimal complete definition

Nothing

Methods

toBody :: a -> RqBody

Headers

(~:) :: FromText a => ResponseHeaders -> HeaderName -> Either String a

(~:?) :: FromText a => ResponseHeaders -> HeaderName -> Either String (Maybe a)

class ToHeaders a where

Minimal complete definition

Nothing

Methods

toHeaders :: a -> [Header]

(=:) :: ToHeader a => HeaderName -> a -> [Header]

hdr :: HeaderName -> ByteString -> [Header] -> [Header]

hdrs :: [Header] -> [Header] -> [Header]

toHeaderText :: ToText a => HeaderName -> a -> [Header]

class ToHeader a where

Minimal complete definition

Nothing

Methods

toHeader :: HeaderName -> a -> [Header]

hHost :: HeaderName

hAMZToken :: HeaderName

hAMZTarget :: HeaderName

hAMZAlgorithm :: HeaderName

hAMZCredential :: HeaderName

hAMZExpires :: HeaderName

hAMZSignedHeaders :: HeaderName

hAMZContentSHA256 :: HeaderName

hAMZAuth :: HeaderName

hAMZDate :: HeaderName

hMetaPrefix :: HeaderName

Path

class ToPath a where

Minimal complete definition

Nothing

Methods

toPath :: a -> Text

Query

class ToQuery a where

Minimal complete definition

Nothing

Methods

toQuery :: a -> Query

renderQuery :: Query -> ByteString

data Query

valuesOf :: Traversal' Query (Maybe ByteString)

(=?) :: ToQuery a => ByteString -> a -> Query

pair :: ToQuery a => ByteString -> a -> Query -> Query

toQueryList :: (IsList a, ToQuery (Item a)) => ByteString -> a -> Query

URI

collapsePath :: ByteString -> ByteString

XML

FromXML

class FromXML a where

Methods

parseXML :: [Node] -> Either String a

decodeXML :: LazyByteString -> Either String [Node]

parseXMLText :: FromText a => String -> [Node] -> Either String a

childNodes :: Text -> Node -> Maybe [Node]

findElement :: Text -> [Node] -> Either String [Node]

withContent :: String -> [Node] -> Either String (Maybe Text)

withElement :: Text -> ([Node] -> Either String a) -> [Node] -> Either String a

localName :: Node -> Maybe Text

(.@) :: FromXML a => [Node] -> Text -> Either String a

(.@?) :: FromXML a => [Node] -> Text -> Either String (Maybe a)

(.!@) :: Either String (Maybe a) -> a -> Either String a

ToXML

class ToXML a where

Minimal complete definition

Nothing

Methods

toXML :: a -> [Node]

class ToXMLRoot a where

Methods

toXMLRoot :: a -> Maybe Element

encodeXML :: ToXMLRoot a => a -> LazyByteString

toXMLText :: ToText a => a -> [Node]

namespaced :: Text -> Text -> [Node] -> Maybe Element

element :: Name -> [Node] -> Element

nodes :: Name -> [Node] -> [Node]

(=@) :: ToXML a => Name -> a -> Node

extractRoot :: Text -> [Node] -> Maybe Element

unsafeToXML :: (Show a, ToXML a) => a -> Node

Caution: This is for use with types which are flattened in AWS service model terminology.

It is applied by the generator/templating in safe contexts only.

JSON

FromJSON

class FromJSON a where

Minimal complete definition

Nothing

Methods

parseJSON :: Value -> Parser a

parseJSONText :: FromText a => String -> Value -> Parser a

eitherDecode' :: FromJSON a => ByteString -> Either String a

Parser a

withObject :: String -> (Object -> Parser a) -> Value -> Parser a

(.:) :: FromJSON a => Object -> Text -> Parser a

(.:?) :: FromJSON a => Object -> Text -> Parser (Maybe a)

(.!=) :: Parser (Maybe a) -> a -> Parser a

Either String a

(.:>) :: FromJSON a => Object -> Text -> Either String a

(.:?>) :: FromJSON a => Object -> Text -> Either String (Maybe a)

ToJSON

class ToJSON a where

Minimal complete definition

Nothing

Methods

toJSON :: a -> Value

toJSONText :: ToText a => a -> Value

type Object = HashMap Text Value

object :: [Pair] -> Value

(.=) :: ToJSON a => Text -> a -> Pair

Collections

newtype List e a

Constructors

List

Fields

list :: [a]

newtype List1 e a

Constructors

List1

Fields

list1 :: NonEmpty a

_List :: (Coercible a b, Coercible b a) => Iso' (List e a) [b]

_List1 :: (Coercible a b, Coercible b a) => Iso' (List1 e a) (NonEmpty b)

fromList1 :: List1 e a -> List e a

toList1 :: List e a -> Either String (List1 e a)

newtype Map k v

Constructors

Map

Fields

fromMap :: HashMap k v

_Map :: (Coercible a b, Coercible b a) => Iso' (Map k a) (HashMap k b)

(~::) :: ResponseHeaders -> CI Text -> Either String (Map (CI Text) Text)

newtype EMap e i j k v

Constructors

EMap

Fields

fromEMap :: HashMap k v

_EMap :: (Coercible a b, Coercible b a) => Iso' (EMap e i j k a) (HashMap k b)