Network.AWS.Data.Body

Description

Documentation

newtype RsBody

A streaming, exception safe response body.

Constructors

RsBody

Fields

_streamBody :: ResumableSource (ResourceT IO) ByteString

fuseStream :: RsBody -> Conduit ByteString (ResourceT IO) ByteString -> RsBody

newtype ChunkSize

Specifies the transmitted size of the 'Transfer-Encoding' chunks.

See: defaultChunk.

Constructors

ChunkSize Int

defaultChunkSize :: ChunkSize

The default chunk size of 128 KB. The minimum chunk size accepted by AWS is 8 KB, unless the entirety of the request is below this threshold.

A chunk size of 64 KB or higher is recommended for performance reasons.

data ChunkedBody

An opaque request body which will be transmitted via Transfer-Encoding: chunked.

Invariant: Only services that support chunked encoding can accept a ChunkedBody. (Currently S3.) This is enforced by the type signatures emitted by the generator.

Constructors

ChunkedBody

Fields

_chunkedSize :: !ChunkSize

_chunkedLength :: !Integer

_chunkedBody :: Source (ResourceT IO) ByteString

chunkedLength :: Lens' ChunkedBody Integer

fuseChunks :: ChunkedBody -> Conduit ByteString (ResourceT IO) ByteString -> ChunkedBody

fullChunks :: ChunkedBody -> Integer

remainderBytes :: ChunkedBody -> Maybe Integer

data HashedBody

An opaque request body containing a SHA256 hash.

Constructors

HashedStream (Digest SHA256) !Integer (Source (ResourceT IO) ByteString)

HashedBytes (Digest SHA256) ByteString

sha256Base16 :: HashedBody -> ByteString

data RqBody

Invariant: only services that support _both_ standard and chunked signing expose RqBody as a parameter.

Constructors

Chunked ChunkedBody

Hashed HashedBody

md5Base64 :: RqBody -> Maybe ByteString

isStreaming :: RqBody -> Bool

toRequestBody :: RqBody -> RequestBody

contentLength :: RqBody -> Integer

class ToHashedBody a where

Anything that can be safely converted to a HashedBody.

Methods

toHashed :: a -> HashedBody

Convert a value to a hashed request body.

class ToBody a where

Anything that can be converted to a streaming request Body.

Minimal complete definition

Nothing

Methods

toBody :: a -> RqBody

Convert a value to a request body.

_Body :: ToBody a => AReview RqBody a