Network.AWS.Types

Authentication

Credentials

newtype AccessKey

Access key credential.

Constructors

AccessKey ByteString

newtype SecretKey

Secret key credential.

Constructors

SecretKey ByteString

newtype SecurityToken

A security token used by STS to temporarily authorise access to an AWS resource.

Constructors

SecurityToken ByteString

Environment

data AuthEnv

The authorisation environment.

Constructors

AuthEnv

Fields

_authAccess :: !AccessKey

_authSecret :: !SecretKey

_authToken :: Maybe SecurityToken

_authExpiry :: Maybe UTCTime

data Auth

An authorisation environment containing AWS credentials, and potentially a reference which can be refreshed out-of-band as temporary credentials expire.

Constructors

Ref ThreadId (IORef AuthEnv)

Auth AuthEnv

withAuth :: MonadIO m => Auth -> (AuthEnv -> m a) -> m a

Services

class (AWSSigner (Sg a), Show (Er a)) => AWSService a where

The properties (such as endpoint) for a service, as well as it's associated signing algorithm and error types.

Associated Types

type Sg a :: *

Signing algorithm supported by the service.

type Er a :: *

The general service error.

Methods

service :: Service a

type Abbrev = Text

Abbreviated service name.

data Service a

Attributes specific to an AWS service.

Constructors

Service

Fields

_svcAbbrev :: !Text

_svcPrefix :: !ByteString

_svcVersion :: !ByteString

_svcTargetPrefix :: Maybe ByteString

_svcJSONVersion :: Maybe ByteString

_svcHandle :: Status -> Maybe (LazyByteString -> ServiceError (Er a))

_svcRetry :: Retry a

serviceOf :: AWSService (Sv a) => Request a -> Service (Sv a)

Retries

data Retry a

Constants and predicates used to create a RetryPolicy.

Constructors

Exponential

Fields

_retryBase :: !Double

_retryGrowth :: !Int

_retryAttempts :: !Int

_retryCheck :: Status -> Er a -> Bool

Endpoints

data Endpoint

Constructors

Endpoint

Fields

_endpointHost :: ByteString

_endpointScope :: ByteString

endpoint :: Service a -> Region -> Endpoint

Determine the full host address and credential scope for a Service within the specified Region.

Errors

data ServiceError a

An error type representing the subset of errors that can be directly attributed to this library.

Constructors

HttpError HttpException

SerializerError Abbrev String

ServiceError Abbrev Status a

Errors [ServiceError a]

_HttpError :: forall a. Prism' (ServiceError a) HttpException

_SerializerError :: forall a. Prism' (ServiceError a) (Abbrev, String)

_ServiceError :: forall a. Prism' (ServiceError a) (Abbrev, Status, a)

_Errors :: forall a. Prism' (ServiceError a) [ServiceError a]

Signing

class AWSSigner v where

Methods

signed :: (AWSService (Sv a), v ~ Sg (Sv a)) => AuthEnv -> Region -> Request a -> UTCTime -> Signed a v

class AWSPresigner v where

Methods

presigned :: (AWSService (Sv a), v ~ Sg (Sv a)) => AuthEnv -> Region -> Request a -> UTCTime -> Integer -> Signed a v

data Signed a v where

A signed ClientRequest and associated metadata specific to the signing algorithm that was used.

Constructors

Signed :: ToBuilder (Meta v) => Meta v -> ClientRequest -> Signed a v

Fields

_sgMeta :: Meta v

_sgRequest :: ClientRequest

data family Meta v :: *

Signing metadata data specific to a signing algorithm.

Note: this is used for test and debug purposes, or is otherwise ignored.

sgMeta :: Lens' (Signed a v) (Meta v)

sgRequest :: Lens' (Signed a v) ClientRequest

Requests

class (AWSService (Sv a), AWSSigner (Sg (Sv a))) => AWSRequest a where

Specify how a request can be de/serialised.

Associated Types

type Sv a :: *

The service definition for a request.

type Rs a :: *

The successful, expected response associated with a request.

Methods

request :: a -> Request a

response :: MonadResource m => Logger -> Request a -> Either HttpException ClientResponse -> m (Response' a)

class AWSRequest a => AWSPager a where

Specify how an AWSRequest and it's associated Rs response can generate a subsequent request, if available.

Methods

page :: a -> Rs a -> Maybe a

data Request a

An unsigned request.

Constructors

Request

Fields

_rqMethod :: !StdMethod

_rqPath :: !ByteString

_rqQuery :: Query

_rqHeaders :: [Header]

_rqBody :: RqBody

rqMethod :: forall a a. Lens (Request a) (Request a) StdMethod StdMethod

rqHeaders :: forall a a. Lens (Request a) (Request a) [Header] [Header]

rqPath :: forall a a. Lens (Request a) (Request a) ByteString ByteString

rqQuery :: forall a a. Lens (Request a) (Request a) Query Query

rqBody :: forall a a. Lens (Request a) (Request a) RqBody RqBody

Responses

type Response a = Either (ServiceError (Er (Sv a))) (Rs a)

An alias for the common response Either containing a service error in the Left case, or the expected response in the Right.

type Response' a = Either (ServiceError (Er (Sv a))) (Status, Rs a)

data Empty

Constructors

Empty

Logging

data LogLevel

Constructors

Info
Informational messages supplied by the user, not used by the library.

Debug
Info level + debug messages + non-streaming response bodies.

Trace
Debug level + potentially sensitive signing metadata.

type Logger = LogLevel -> Builder -> IO ()

Regions

data Region

The sum of available AWS regions.

Constructors

Ireland
Europe / eu-west-1

Frankfurt
Europe / eu-central-1

Tokyo
Asia Pacific / ap-northeast-1

Singapore
Asia Pacific / ap-southeast-1

Sydney
Asia Pacific / ap-southeast-2

Beijing
China / cn-north-1

NorthVirginia
US / us-east-1

NorthCalifornia
US / us-west-1

Oregon
US / us-west-2

GovCloud
AWS GovCloud / us-gov-west-1

GovCloudFIPS
AWS GovCloud (FIPS 140-2) S3 Only / fips-us-gov-west-1

SaoPaulo
South America / sa-east-1

Query Actions

newtype Action

A service's query action.

Constructors

Action Text

Convenience

type ClientRequest = Request

A convenience alias to avoid type ambiguity.

type ClientResponse = Response ResponseBody

A convenience alias encapsulating the common Response.

type ResponseBody = ResumableSource (ResourceT IO) ByteString

A convenience alias encapsulating the common Response body.

clientRequest :: ClientRequest

Construct a ClientRequest using common parameters such as TLS and prevent throwing errors when receiving erroneous status codes in respones.