Network.AWS.Prelude

Re-exported

Primitives

data ByteString :: *

A space-efficient representation of a Word8 vector, supporting many efficient operations.

A ByteString contains 8-bit bytes, or by using the operations from Data.ByteString.Char8 it can be interpreted as containing 8-bit characters.

data CI s :: * -> *

type ClientRequest = Request

A convenience alias to avoid type ambiguity.

type ClientResponse = Response ResponseBody

A convenience alias encapsulating the common Response.

class (Typeable * e, Show e) => Exception e

Any type that you wish to throw or catch as an exception must be an instance of the Exception class. The simplest case is a new exception type directly below the root:

data MyException = ThisException | ThatException
    deriving (Show, Typeable)

instance Exception MyException

The default method definitions in the Exception class do what we need in this case. You can now throw and catch ThisException and ThatException as exceptions:

*Main> throw ThisException `catch` \e -> putStrLn ("Caught " ++ show (e :: MyException))
Caught ThisException

In more complicated examples, you may wish to define a whole hierarchy of exceptions:

---------------------------------------------------------------------
-- Make the root exception type for all the exceptions in a compiler

data SomeCompilerException = forall e . Exception e => SomeCompilerException e
    deriving Typeable

instance Show SomeCompilerException where
    show (SomeCompilerException e) = show e

instance Exception SomeCompilerException

compilerExceptionToException :: Exception e => e -> SomeException
compilerExceptionToException = toException . SomeCompilerException

compilerExceptionFromException :: Exception e => SomeException -> Maybe e
compilerExceptionFromException x = do
    SomeCompilerException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make a subhierarchy for exceptions in the frontend of the compiler

data SomeFrontendException = forall e . Exception e => SomeFrontendException e
    deriving Typeable

instance Show SomeFrontendException where
    show (SomeFrontendException e) = show e

instance Exception SomeFrontendException where
    toException = compilerExceptionToException
    fromException = compilerExceptionFromException

frontendExceptionToException :: Exception e => e -> SomeException
frontendExceptionToException = toException . SomeFrontendException

frontendExceptionFromException :: Exception e => SomeException -> Maybe e
frontendExceptionFromException x = do
    SomeFrontendException a <- fromException x
    cast a

---------------------------------------------------------------------
-- Make an exception type for a particular frontend compiler exception

data MismatchedParentheses = MismatchedParentheses
    deriving (Typeable, Show)

instance Exception MismatchedParentheses where
    toException   = frontendExceptionToException
    fromException = frontendExceptionFromException

We can now catch a MismatchedParentheses exception as MismatchedParentheses, SomeFrontendException or SomeCompilerException, but not other types, e.g. IOException:

*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: MismatchedParentheses))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: SomeFrontendException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: SomeCompilerException))
Caught MismatchedParentheses
*Main> throw MismatchedParentheses catch e -> putStrLn ("Caught " ++ show (e :: IOException))
*** Exception: MismatchedParentheses

data HashMap k v :: * -> * -> *

data HttpException :: *

data Natural :: *

Type representing arbitrary-precision non-negative integers.

Operations whose result would be negative throw (Underflow :: ArithException).

Since: 4.8.0.0

data NonEmpty a :: * -> *

Constructors

a :| [a]

data RequestBody :: *

type Response a = Either (ServiceError (Er (Sv a))) (Rs a)

An alias for the common response Either containing a service error in the Left case, or the expected response in the Right.

data Text :: *

Classes

class AWSRequest a => AWSPager a where

Specify how an AWSRequest and it's associated Rs response can generate a subsequent request, if available.

Methods

page :: a -> Rs a -> Maybe a

class (AWSService (Sv a), AWSSigner (Sg (Sv a))) => AWSRequest a where

Specify how a request can be de/serialised.

Associated Types

type Sv a :: *

The service definition for a request.

type Rs a :: *

The successful, expected response associated with a request.

Methods

request :: a -> Request a

response :: MonadResource m => Logger -> Request a -> Either HttpException ClientResponse -> m (Response' a)

class (AWSSigner (Sg a), Show (Er a)) => AWSService a where

The properties (such as endpoint) for a service, as well as it's associated signing algorithm and error types.

Associated Types

type Sg a :: *

Signing algorithm supported by the service.

type Er a :: *

The general service error.

Methods

service :: Service a

class Generic a

Representable types of kind *. This class is derivable in GHC with the DeriveGeneric flag on.

Minimal complete definition

from, to

class IsString a where

Class for string-like datastructures; used by the overloaded string extension (-XOverloadedStrings in GHC).

Methods

fromString :: String -> a

class Semigroup a

Retries

data Retry a

Constants and predicates used to create a RetryPolicy.

Constructors

Exponential

Fields

_retryBase :: !Double

_retryGrowth :: !Int

_retryAttempts :: !Int

_retryCheck :: Status -> Er a -> Bool

Errors

class AWSError a where

Methods

awsError :: a -> ServiceError String

class AWSErrorCode a where

Methods

awsErrorCode :: a -> ErrorCode

data ServiceError a

An error type representing the subset of errors that can be directly attributed to this library.

Constructors

HttpError HttpException

SerializerError Abbrev String

ServiceError Abbrev Status a

Errors [ServiceError a]

data RESTError

restError :: FromXML (Er a) => (Status -> Bool) -> Service a -> Status -> Maybe (ByteString -> ServiceError (Er a))

data JSONError

jsonError :: FromJSON (Er a) => (Status -> Bool) -> Service a -> Status -> Maybe (ByteString -> ServiceError (Er a))

statusSuccess :: Status -> Bool

Shared

data Empty

Constructors

Empty

data Service a

Attributes specific to an AWS service.

Constructors

Service

Fields

_svcAbbrev :: !Text

_svcPrefix :: !ByteString

_svcVersion :: !ByteString

_svcTargetPrefix :: Maybe ByteString

_svcJSONVersion :: Maybe ByteString

_svcHandle :: Status -> Maybe (LazyByteString -> ServiceError (Er a))

_svcRetry :: Retry a

HTTP

data StdMethod :: *

Constructors

GET

POST

HEAD

PUT

DELETE

TRACE

CONNECT

OPTIONS

PATCH

data Status :: *

Constructors

Status

Fields

statusCode :: Int

statusMessage :: ByteString