Network.AWS.Types

Description

Authentication

Credentials

newtype AccessKey

An access key ID.

For example: AKIAIOSFODNN7EXAMPLE

See: Understanding and Getting Your Security Credentials.

Constructors

AccessKey ByteString

newtype SecretKey

Secret access key credential.

For example: wJalrXUtnFEMIK7MDENGbPxRfiCYEXAMPLEKE

See: Understanding and Getting Your Security Credentials.

Constructors

SecretKey ByteString

newtype SessionToken

A session token used by STS to temporarily authorise access to an AWS resource.

See: Temporary Security Credentials.

Constructors

SessionToken ByteString

Environment

data Auth

An authorisation environment containing AWS credentials, and potentially a reference which can be refreshed out-of-band as temporary credentials expire.

Constructors

Ref ThreadId (IORef AuthEnv)

Auth AuthEnv

withAuth :: MonadIO m => Auth -> (AuthEnv -> m a) -> m a

data AuthEnv

The AuthN/AuthZ credential environment.

Constructors

AuthEnv

Fields

_authAccess :: !AccessKey

_authSecret :: !(Sensitive SecretKey)

_authToken :: Maybe (Sensitive SessionToken)

_authExpiry :: Maybe ISO8601

accessKeyId :: Lens' AuthEnv AccessKey

The access key ID that identifies the temporary security credentials.

secretAccessKey :: Lens' AuthEnv SecretKey

The secret access key that can be used to sign requests.

sessionToken :: Lens' AuthEnv (Maybe SessionToken)

The token that users must pass to the service API to use the temporary credentials.

expiration :: Lens' AuthEnv (Maybe UTCTime)

The date on which the current credentials expire.

Logging

data LogLevel

Constructors

Info
Info messages supplied by the user - this level is not emitted by the library.

Error
Error messages only.

Debug
Useful debug information + info + error levels.

Trace
Includes potentially sensitive signing metadata, and non-streaming response bodies.

type Logger = LogLevel -> Builder -> IO ()

A function threaded through various request and serialisation routines to log informational and debug messages.

Signing

type Algorithm a = Request a -> AuthEnv -> Region -> UTCTime -> Signed a

data Meta where

Signing algorithm specific metadata.

Constructors

Meta :: ToLog a => a -> Meta

data Signer

Constructors

Signer

Fields

sgSign :: forall a. Algorithm a

sgPresign :: forall a. Seconds -> Algorithm a

data Signed a

A signed ClientRequest and associated metadata specific to the signing algorithm, tagged with the initial request type to be able to obtain the associated response, 'Rs a'.

Constructors

Signed

Fields

sgMeta :: !Meta

sgRequest :: !ClientRequest

Service

data Abbrev

Abbreviated service name.

data Service

Attributes and functions specific to an AWS service.

Constructors

Service

Fields

_svcAbbrev :: !Abbrev

_svcSigner :: !Signer

_svcPrefix :: !ByteString

_svcVersion :: !ByteString

_svcEndpoint :: !(Region -> Endpoint)

_svcTimeout :: !(Maybe Seconds)

_svcCheck :: !(Status -> Bool)

_svcError :: !(Status -> [Header] -> LazyByteString -> Error)

_svcRetry :: !Retry

serviceSigner :: Lens' Service Signer

serviceEndpoint :: Setter' Service Endpoint

serviceTimeout :: Lens' Service (Maybe Seconds)

serviceCheck :: Lens' Service (Status -> Bool)

serviceRetry :: Lens' Service Retry

Requests

class AWSRequest a where

Specify how a request can be de/serialised.

Minimal complete definition

request, response

Associated Types

type Rs a :: *

The successful, expected response associated with a request.

Methods

request :: a -> Request a

response :: MonadResource m => Logger -> Service -> Proxy a -> ClientResponse -> m (Response a)

data Request a

An unsigned request.

Constructors

Request

Fields

_rqService :: !Service

_rqMethod :: !StdMethod

_rqPath :: !RawPath

_rqQuery :: !QueryString

_rqHeaders :: ![Header]

_rqBody :: !RqBody

rqService :: Lens' (Request a) Service

rqMethod :: Lens' (Request a) StdMethod

rqHeaders :: Lens' (Request a) [Header]

rqPath :: Lens' (Request a) RawPath

rqQuery :: Lens' (Request a) QueryString

rqBody :: Lens' (Request a) RqBody

rqSign :: Algorithm a

rqPresign :: Seconds -> Algorithm a

Responses

type Response a = (Status, Rs a)

Retries

data Retry

Constants and predicates used to create a RetryPolicy.

Constructors

Exponential

Fields

_retryBase :: !Double

_retryGrowth :: !Int

_retryAttempts :: !Int

_retryCheck :: ServiceError -> Maybe Text

Returns a descriptive name for logging if the request should be retried.

exponentBase :: Lens' Retry Double

exponentGrowth :: Lens' Retry Int

retryAttempts :: Lens' Retry Int

retryCheck :: Lens' Retry (ServiceError -> Maybe Text)

Errors

class AsError a where

Minimal complete definition

_Error

Methods

_Error :: Prism' a Error

A general Amazonka error.

_TransportError :: Prism' a HttpException

An error occured while communicating over HTTP with a remote service.

_SerializeError :: Prism' a SerializeError

A serialisation error occured when attempting to deserialise a response.

_ServiceError :: Prism' a ServiceError

A service specific error returned by the remote service.

data Error

An error type representing errors that can be attributed to this library.

Constructors

TransportError HttpException

SerializeError SerializeError

ServiceError ServiceError

HTTP Errors

data HttpException :: *

Serialize Errors

data SerializeError

Constructors

SerializeError'

Fields

_serializeAbbrev :: !Abbrev

_serializeStatus :: !Status

_serializeBody :: Maybe LazyByteString

The response body, if the response was not streaming.

_serializeMessage :: String

serializeAbbrev :: Lens' SerializeError Abbrev

serializeStatus :: Lens' SerializeError Status

serializeMessage :: Lens' SerializeError String

Service Errors

data ServiceError

Constructors

ServiceError'

Fields

_serviceAbbrev :: !Abbrev

_serviceStatus :: !Status

_serviceHeaders :: [Header]

_serviceCode :: !ErrorCode

_serviceMessage :: Maybe ErrorMessage

_serviceRequestId :: Maybe RequestId

serviceAbbrev :: Lens' ServiceError Abbrev

serviceStatus :: Lens' ServiceError Status

serviceHeaders :: Lens' ServiceError [Header]

serviceCode :: Lens' ServiceError ErrorCode

serviceMessage :: Lens' ServiceError (Maybe ErrorMessage)

serviceRequestId :: Lens' ServiceError (Maybe RequestId)

Error Types

newtype ErrorCode

Constructors

ErrorCode Text

errorCode :: Text -> ErrorCode

Construct an ErrorCode.

newtype ErrorMessage

Constructors

ErrorMessage Text

newtype RequestId

Constructors

RequestId Text

Regions

data Region

The available AWS regions.

Constructors

NorthVirginia
US East ('us-east-1').

Ohio
US East ('us-east-2').

NorthCalifornia
US West ('us-west-1').

Oregon
US West ('us-west-2').

Montreal
Canada ('ca-central-1').

Tokyo
Asia Pacific ('ap-northeast-1').

Seoul
Asia Pacific ('ap-northeast-2').

Mumbai
Asia Pacific ('ap-south-1').

Singapore
Asia Pacific ('ap-southeast-1').

Sydney
Asia Pacific ('ap-southeast-2').

SaoPaulo
South America ('sa-east-1').

Ireland
EU ('eu-west-1').

London
EU ('eu-west-2').

Frankfurt
EU ('eu-central-1').

GovCloud
US GovCloud ('us-gov-west-1').

GovCloudFIPS
US GovCloud FIPS (S3 Only, 'fips-us-gov-west-1').

Beijing
China ('cn-north-1').

Endpoints

data Endpoint

Constructors

Endpoint

Fields

_endpointHost :: ByteString

_endpointSecure :: !Bool

_endpointPort :: !Int

_endpointScope :: ByteString

endpointHost :: Lens' Endpoint ByteString

endpointPort :: Lens' Endpoint Int

endpointSecure :: Lens' Endpoint Bool

endpointScope :: Lens' Endpoint ByteString

HTTP

type ClientRequest = Request

A convenience alias to avoid type ambiguity.

type ClientResponse = Response ResponseBody

A convenience alias encapsulating the common Response.

type ResponseBody = ResumableSource (ResourceT IO) ByteString

A convenience alias encapsulating the common Response body.

clientRequest :: Endpoint -> Maybe Seconds -> ClientRequest

Construct a ClientRequest using common parameters such as TLS and prevent throwing errors when receiving erroneous status codes in respones.

Seconds

newtype Seconds

An integral value representing seconds.

Constructors

Seconds Int

seconds :: Seconds -> Int

microseconds :: Seconds -> Int

Isomorphisms

_Coerce :: (Coercible a b, Coercible b a) => Iso' a b

_Default :: Monoid a => Iso' (Maybe a) a

Invalid Iso, should be a Prism but exists for ease of composition with the current 'Lens . Iso' chaining to hide internal types from the user.